# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "APIResponse_AnnotationTaskResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnnotationTaskResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[AnnotationTaskResponse]",
        "type": "object"
      },
...
      "APIResponse_JobAssignmentResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobAssignmentResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[JobAssignmentResponse]",
        "type": "object"
      },
...
      "APIResponse_NoneType_": {
        "properties": {
          "data": {
            "description": "Endpoint specific payload.",
            "title": "Data",
            "type": "null"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[NoneType]",
        "type": "object"
      },
...
      "APIResponse_TokenResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[TokenResponse]",
        "type": "object"
      },
      "APIResponse_UserResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[UserResponse]",
        "type": "object"
      },
...
      "APIResponse_list_JobAssignmentResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JobAssignmentResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[JobAssignmentResponse]]",
        "type": "object"
      },
...
      "AnnotationTaskAssign": {
        "description": "Payload used to assign a task to an annotator.",
        "properties": {
          "assigned_to": {
            "exclusiveMinimum": 0.0,
            "title": "Assigned To",
            "type": "integer"
          }
        },
        "required": [
          "assigned_to"
        ],
        "title": "AnnotationTaskAssign",
        "type": "object"
      },
...
      "AnnotationTaskResponse": {
        "description": "Annotation task representation returned by the API.",
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Record Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "input_text": {
            "description": "Record, sentence or image reference to annotate.",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Input Text",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "input_text",
          "id",
          "job_id",
          "dataset_record_id",
          "assigned_to",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AnnotationTaskResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobAssignmentCreate": {
        "description": "Payload for creating an assignment.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "status": {
            "default": "assigned",
            "maxLength": 30,
            "title": "Status",
            "type": "string"
          },
          "worker_id": {
            "title": "Worker Id",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "worker_id"
        ],
        "title": "JobAssignmentCreate",
        "type": "object"
      },
      "JobAssignmentResponse": {
        "description": "Assignment representation returned by the API.",
        "properties": {
          "assigned_at": {
            "format": "date-time",
            "title": "Assigned At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "status": {
            "default": "assigned",
            "maxLength": 30,
            "title": "Status",
            "type": "string"
          },
          "worker_id": {
            "title": "Worker Id",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "worker_id",
          "id",
          "assigned_at"
        ],
        "title": "JobAssignmentResponse",
        "type": "object"
      },
      "JobAssignmentUpdate": {
        "description": "Payload for updating an assignment.",
        "properties": {
          "status": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "JobAssignmentUpdate",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "Credentials submitted to the login endpoint.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "JWT response returned after successful authentication.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "UserCreate": {
        "description": "Payload used to register a new account.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "default": "annotator",
            "description": "Self-registration supports the dataset_owner and annotator roles. Administrator accounts are created by an administrator.",
            "maxLength": 30,
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserResponse": {
        "description": "Public user representation returned by the API.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "id",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the AI Dataset Labeling Marketplace: datasets, labeling jobs, annotation tasks, annotations, quality review, progress tracking, dataset export and the AI annotation assistant.",
    "title": "AI Dataset Labeling Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/assignments": {
      "get": {
        "description": "Return every assignment for the management roles.",
        "operationId": "get_assignments_api_assignments_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_JobAssignmentResponse__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List job assignments",
        "tags": [
          "Job Assignments"
        ]
      },
      "post": {
        "description": "Create a job assignment for an active annotator.",
        "operationId": "create_assignment_api_assignments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobAssignmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_JobAssignmentResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign a labeling job to an annotator",
        "tags": [
          "Job Assignments"
        ]
      }
    },
    "/api/assignments/job/{job_id}": {
      "get": {
        "description": "Return every assignment created for a job.",
        "operationId": "get_job_assignments_api_assignments_job__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_JobAssignmentResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the assignments of a labeling job",
        "tags": [
          "Job Assignments"
        ]
      }
    },
    "/api/assignments/mine": {
      "get": {
        "description": "Return the assignments belonging to the caller.",
        "operationId": "get_my_assignments_api_assignments_mine_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_JobAssignmentResponse__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the assignments of the authenticated annotator",
        "tags": [
          "Job Assignments"
        ]
      }
    },
    "/api/assignments/{assignment_id}": {
      "delete": {
        "description": "Remove an assignment from a labeling job.",
        "operationId": "delete_assignment_api_assignments__assignment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete an assignment",
        "tags": [
          "Job Assignments"
        ]
      },
      "get": {
        "description": "Return a single assignment by identifier.",
        "operationId": "get_assignment_api_assignments__assignment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_JobAssignmentResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retrieve one assignment",
        "tags": [
          "Job Assignments"
        ]
      },
      "put": {
        "description": "Change the status of an existing assignment.",
        "operationId": "update_assignment_api_assignments__assignment_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobAssignmentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_JobAssignmentResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update the status of an assignment",
        "tags": [
          "Job Assignments"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Validate credentials and issue a signed JWT.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TokenResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authenticate and receive a JWT access token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "Return the profile of the caller, including the role.",
        "operationId": "get_authenticated_user_api_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UserResponse_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Return the authenticated user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Create a new account and return the public profile.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UserResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a new dataset owner or annotator account",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/tasks/{task_id}/assign": {
      "post": {
        "description": "Hand a task over to an annotator.",
        "operationId": "assign_task_api_tasks__task_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationTaskAssign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AnnotationTaskResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign an annotation task to an annotator",
        "tags": [
          "Annotation Tasks"
        ]
      }
    },
...
  }
}